pyproject.toml
[tool.poetry]
name = "sanic-tortoise-user-api"
version = "0.1.0"
description = ""
authors = ["Gustavo Correa <self.gustavocorrea@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
sanic = "^21.9.1"
tortoise-orm = {extras = ["asyncpg"], version = "^0.17.8"}
aerich = "^0.5.8"
sanic-ext = "^21.9.0"
ujson = "^4.2.0"
pydantic = {extras = ["email"], version = "^1.8.2"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
aiosqlite = "^0.17.0"
autoflake = "^1.4"
typing-extensions = "^3.10.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}

[tool.poetry.dev-dependencies]
black = "^21.9b0"
isort = "^5.9.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
